Added authorization roles controls from properties Changed the structure of beans.
