+++ partner/partner-management-service/src/main/resources/application-local.properties
@@ -149,3 +149,63 @@ apikey.status.content-template=apikey-status-content-template
 pms.notifications-schedule.init-delay = 24
 pms.notifications-schedule.fixed-rate=24
 
+
+mosip.role.pms.postdevicedetail=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.putdevicedetail=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.patchdevicedetail=PARTNER_ADMIN
+mosip.role.pms.postdevicedetailsearch=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.postdevicedetaildevicetypesearch=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.postdevicedetailfiltervalues=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.postdevicedetaildevicetypefiltervalues=PARTNER_ADMIN,DEVICE_PROVIDER,FTM_PROVIDER
+mosip.role.pms.postdevicedetaildevicesubtypefiltervalues=PARTNER_ADMIN,FTM_PROVIDER,DEVICE_PROVIDER
+mosip.role.pms.postdeviceprovidermanagementvalidate=ZONAL_ADMIN,ID_AUTHENTICATION,REGISTRATION_PROCESSOR,RESIDENT
+mosip.role.pms.postftpchipdetail=PARTNER_ADMIN,FTM_PROVIDER
+mosip.role.pms.putftpchipdetail=PARTNER_ADMIN,FTM_PROVIDER
+mosip.role.pms.patchftpchipdetail=PARTNER_ADMIN
+mosip.role.pms.postftpchipdetailuploadcertificate=PARTNER_ADMIN,FTM_PROVIDER
+mosip.role.pms.getftpchipdetailgetpartnercertificate=PARTNER_ADMIN,FTM_PROVIDER
+mosip.role.pms.postftpchipdetailsearch=PARTNER_ADMIN,FTM_PROVIDER
+mosip.role.pms.postregistereddevices=PARTNER_ADMIN,FTM_PROVIDER,ZONAL_ADMIN
+mosip.role.pms.postregistereddevicesderegister=PARTNER_ADMIN,FTM_PROVIDER,ZONAL_ADMIN
+mosip.role.pms.postregistereddevicessearch=PARTNER,FTM_PROVIDER,PMS_USER,AUTH_PARTNER,DEVICE_PROVIDER,CREDENTIAL_PARTNER,CREDENTIAL_ISSUANCE,CREATE_SHARE,ID_AUTHENTICATION
+mosip.role.pms.postsecurebiometricinterface=PARTNER_ADMIN,FTM_PROVIDER,DEVICE_PROVIDER
+mosip.role.pms.putsecurebiometricinterface=PARTNER_ADMIN,FTM_PROVIDER,DEVICE_PROVIDER
+mosip.role.pms.patchsecurebiometricinterface=PARTNER_ADMIN
+mosip.role.pms.postsecurebiometricinterfacesearch=PARTNER_ADMIN,FTM_PROVIDER,DEVICE_PROVIDER
+mosip.role.pms.postpartners=PARTNER,PARTNER_ADMIN,AUTH_PARTNER,CREDENTIAL_PARTNER
+mosip.role.pms.patchpartnerspartneridapikeyrequest=PARTNER_ADMIN,AUTH_PARTNER,PARTNER,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_PARTNER
+mosip.role.pms.postpartnerspartneridbioextractorspolicyid=CREATE_SHARE,PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.getpartnerspartneridbioextractorspolicyid=CREATE_SHARE,PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.postpartnerspartneridcredentialtypepolicyid=CREATE_SHARE,PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.getpartnerspartneridcredentialtypepolicies=CREATE_SHARE,PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.postpartnerspartneridcontactadd=CREATE_SHARE,PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.putpartnerspartnerid=PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,AUTH_PARTNER
+mosip.role.pms.getpartnerspartnerid=PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,AUTH_PARTNER,RESIDENT
+mosip.role.pms.getpartnerspartneridapikeyrequest=PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,AUTH_PARTNER
+mosip.role.pms.getpartnerspartneridapikeyapikeyreqid=PARTNER,PARTNER_ADMIN,ONLINE_VERIFICATION_PARTNER,CREDENTIAL_ISSUANCE,AUTH_PARTNER
+mosip.role.pms.postpartnerscertificatecaupload=PARTNER_ADMIN,FTM_PROVIDER,AUTH_PARTNER,PMS_USER,ONLINE_VERIFICATION_PARTNER,PARTNERMANAGER
+mosip.role.pms.postpartnerscertificateupload=PARTNER_ADMIN,PARTNER_ADMIN,FTM_PROVIDER,AUTH_PARTNER,PMS_USER,ONLINE_VERIFICATION_PARTNER,PARTNERMANAGER,ID_AUTHENTICATION,CREDENTIAL_ISSUANCE,CREDENTIAL_PARTNER
+mosip.role.pms.getpartnerspartneridcertificate=PARTNER_ADMIN,FTM_PROVIDER,PMS_USER,AUTH_PARTNER,DEVICE_PROVIDER,CREDENTIAL_PARTNER,CREDENTIAL_ISSUANCE,CREATE_SHARE,ID_AUTHENTICATION,ONLINE_VERIFICATION_PARTNER
+mosip.role.pms.postpartnerssearch=PARTNER_ADMIN,FTM_PROVIDER,PMS_USER,AUTH_PARTNER,DEVICE_PROVIDER,CREDENTIAL_PARTNER,CREDENTIAL_ISSUANCE,CREATE_SHARE,ID_AUTHENTICATION,ONLINE_VERIFICATION_PARTNER
+mosip.role.pms.postpartnerspartnertypesearch=PARTNER_ADMIN,FTM_PROVIDER,PMS_USER,AUTH_PARTNER,DEVICE_PROVIDER,CREDENTIAL_PARTNER,CREDENTIAL_ISSUANCE,CREATE_SHARE,ID_AUTHENTICATION,ONLINE_VERIFICATION_PARTNER
+mosip.role.pms.postpartnersfiltervalues=PARTNER_ADMIN,PARTNER,ONLINE_VERIFICATION_PARTNER
+mosip.role.pms.postpartnersapikeyrequestfiltervalues=PARTNER_ADMIN,FTM_PROVIDER,PARTNER,AUTH_PARTNER,CREDENTIAL_PARTNER
+mosip.role.pms.postpartnersapikeyrequestsearch=PARTNER_ADMIN,FTM_PROVIDER,PARTNER,AUTH_PARTNER,CREDENTIAL_PARTNER,DEVICE_PROVIDER,PMS_USER,ONLINE_VERIFICATION_PARTNER
+mosip.role.pms.postpartnersapikeysearch=PARTNER_ADMIN,FTM_PROVIDER,PARTNER,AUTH_PARTNER,CREDENTIAL_PARTNER,DEVICE_PROVIDER,ONLINE_VERIFICATION_PARTNER,PMS_USER
+mosip.role.pms.putpartnerspartneridapikeypolicies=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.patchpartnerspartnerid=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.patchpartnerspartneridapikey=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.getpartnerspartneridapikeymisplicensekey=PARTNERMANAGER,PARTNER_ADMIN,ID_AUTHENTICATION,CREDENTIAL_ISSUANCE
+mosip.role.pms.patchpartnersapikey=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.getpartners=REGISTRATION_PROCESSOR,RESIDENT,PARTNER_ADMIN,ID_AUTHENTICATION,CREDENTIAL_ISSUANCE,PARTNERMANAGER
+mosip.role.pms.getpartnerspartneridapikey=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.getpartnersapikey=PARTNERMANAGER,PARTNER_ADMIN
+mosip.role.pms.getpartnersapikeykey=PARTNER_ADMIN,PARTNERMANAGER
+mosip.role.pms.postmisps=PARTNER_ADMIN,MISP_PARTNER,MISP,PARTNERMANAGER
+mosip.role.pms.putmisps=PARTNER_ADMIN,MISP_PARTNER,MISP,PARTNERMANAGER
+mosip.role.pms.getmisps=PARTNER_ADMIN,MISP_PARTNER,MISP,PARTNERMANAGER
+mosip.role.pms.getmispsmispidlicensekey=PARTNER_ADMIN,MISP_PARTNER,MISP,PARTNERMANAGER
+
+
+
+
