+++ partner/policy-management-service/src/main/resources/application-local.properties
@@ -52,3 +52,21 @@ pmp.api.version=v1
 server.port=9101
 server.servlet.context-path=/partnermanagement/${pmp.api.version}/policies
 partner.search.maximum.rows=5
+
+mosip.role.pms.postpoliciesgroupnew=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.putpoliciesgrouppolicygroupid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.postpolicies=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.postpoliciespolicyidgrouppublish=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.putpoliciespolicyid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.patchpoliciespolicyidgrouppolicygroupid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpolicies=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciespolicyid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciesapikey=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciespolicyidpartnerpartnerid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciesgrouppolicygroupid=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciesgroupall=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.postpoliciesgroupsearch=PARTNER_ADMIN,AUTH_PARTNER,CREDENTIAL_PARTNER,PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.postpoliciessearch=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.getpoliciesconfigkey=PARTNER_ADMIN,POLICYMANAGER
+mosip.role.pms.postpoliciesgroupfiltervalues=PARTNER_ADMIN,PMS_USER,AUTH_PARTNER,CREDENTIAL_PARTNER,POLICYMANAGER,PARTNER
+mosip.role.pms.postpoliciesfiltervalues=PARTNER_ADMIN,PMS_USER,AUTH_PARTNER,CREDENTIAL_PARTNER,POLICYMANAGER,PARTNER
